comments
"""

This blog is based on the original house an article from issue 68 of the
Magpi Magazine which you can buy here
https://www.raspberrypi.org/magpi/issues/68/


link download for free from here.

https://www.raspberrypi.org/magpi-issues/MagPi68.pdf


The article is on page 54.  It was written by an man called Steve Martin who you
can access on Twitter here:

http://twitter.com/0x90_Bug

Without digging too deep it seems as though he helps
run/ attends the Cotswold Jam. This is an excellent tutorial in how to get started
with Python and Minecraft. I have slightly modified and extended the code to work with
the Micro:bit. This version creates a 3-4 storey version which creates 2 sides of a "road".
It is buggy as the right handside faces the wrong way.  

Extended by @ChrisPenn84

"""